% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{age_depth_model}
\alias{age_depth_model}
\title{Create age depth models}
\usage{
age_depth_model(.data, depth, age, age_min = NA_real_, age_max = NA_real_,
  interpolate_age = trans_interpolate,
  extrapolate_age_below = ~trans_average(.x, .y, x0 = last, y0 = last),
  extrapolate_age_above = ~trans_average(.x, .y, x0 = first, y0 = first),
  interpolate_age_err = trans_exact, extrapolate_age_err_below = trans_na,
  extrapolate_age_err_above = trans_na)
}
\arguments{
\item{.data}{A data frame}

\item{depth, age, age_min, age_max}{Expressions evaluated in \code{.data} that provide the
known depths, known ages, and error information if available. These expressions
are evaluated like they are within \link[dplyr]{mutate}.}

\item{interpolate_age, extrapolate_age_below, extrapolate_age_above}{These arguments
provide the rules for interpolating and extrapolating ages based on depths.}

\item{interpolate_age_err, extrapolate_age_err_below, extrapolate_age_err_above}{These arguments
provide the rules for interpolating and extrapolating age min and max values
based on depths.}
}
\value{
An age depth model object.
}
\description{
Create age depth models
}
