% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_axis.R
\name{as_sec_axis}
\alias{as_sec_axis}
\alias{as_sec_axis.age_depth_model}
\title{Use Age Depth Models as a Second ggplot Axis}
\usage{
as_sec_axis(x, ...)

\method{as_sec_axis}{age_depth_model}(x, ..., primary = c("depth", "age"))
}
\arguments{
\item{x}{An \link{age_depth_model}}

\item{...}{Passed to \link[ggplot2]{sec_axis}}

\item{primary}{Specify the primary axis as 'age' or 'depth'}
}
\value{
A ggplot2 \link[ggplot2]{sec_axis} for use in \link[ggplot2]{scale_x_continuous},
  \link[ggplot2]{scale_y_continuous}, or their reverse variants.
}
\description{
Use Age Depth Models as a Second ggplot Axis
}
\examples{
library(ggplot2)
alta_lake_adm <- age_depth_model(
  alta_lake_210Pb_ages, depth = depth_cm, age = age_year_ad
)

ggplot(alta_lake_210Pb_ages, aes(y = depth_cm, x = age_year_ad)) +
  geom_line() +
  geom_point() +
  scale_y_reverse(sec.axis = as_sec_axis(alta_lake_adm))

}
