% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{as_trans_factory}
\alias{as_trans_factory}
\alias{validate_trans_factory}
\alias{validate_trans}
\title{Coerce and validate transforms and functions that produce them}
\usage{
as_trans_factory(factory, env = parent.frame())

validate_trans_factory(factory, x = 1:3, y = 1:3)

validate_trans(trans, x = 1:3, y = 1:3)
}
\arguments{
\item{factory}{A function that produces a transform object}

\item{env}{The calling environment, for transform factories that are calls or
rlang lambda-style functions.}

\item{x}{The test x data}

\item{y}{The test y data}

\item{trans}{A transform object}
}
\value{
The input, invisibly.
}
\description{
Coerce and validate transforms and functions that produce them
}
\examples{
as_trans_factory(age_depth_interpolate)

}
