% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_prcomp.R
\name{autoplot.nested_prcomp}
\alias{autoplot.nested_prcomp}
\title{A ggplot2-based PCA biplot}
\usage{
\method{autoplot}{nested_prcomp}(object, mapping = NULL, which = c("PC1",
  "PC2"), ..., label = .data$row_number, score_geom = ggplot2::geom_point(),
  score_label_geom = ggrepel::geom_text_repel(ggplot2::aes(label =
  .data$auto_label), alpha = 0.5),
  var_geom = list(ggplot2::geom_segment(ggplot2::aes(xend = 0, yend = 0),
  color = "red"), ggplot2::geom_point(color = "red")),
  var_label_geom = ggplot2::geom_label(ggplot2::aes(label = .data$variable),
  color = "red", hjust = "outward", vjust = "outward", fill = NA, label.size =
  NA), var_buffer_geom = ggplot2::geom_blank(),
  default_theme = ggplot2::theme_bw(), env = parent.frame())
}
\arguments{
\item{object}{A \link{nested_prcomp} object}

\item{mapping}{A mapping, created with \link[ggplot2]{aes}.}

\item{which}{Which components to plot}

\item{...}{Passed to \link[ggplot2]{facet_wrap} if there are grouping variables, otherwise ignored}

\item{label}{An expression evaluated within unnest(object, qualifiers, scores) that defines the label for the object.}

\item{score_geom, score_label_geom, var_geom, var_label_geom, var_buffer_geom}{Override these to customize the appearence of
each layer, or use NULL to suppress.}

\item{default_theme}{Use NULL to ignore the default theme for this plot, \link[ggplot2]{theme_bw}.}

\item{env}{Plot environment}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
A ggplot2-based PCA biplot
}
