% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat.R
\name{facet_abundanceh}
\alias{facet_abundanceh}
\alias{facet_abundance}
\title{Facet for relative abundance data}
\usage{
facet_abundanceh(taxon, grouping = NULL, rotate_facet_labels = 45)

facet_abundance(taxon, grouping = NULL, rotate_facet_labels = 0)
}
\arguments{
\item{taxon}{A call to \link[ggplot2]{vars}, defining the column that identifies the taxon.}

\item{grouping}{A call to \link[ggplot2]{vars}, identifying additional grouping columns}

\item{rotate_facet_labels}{Facet label rotation (degrees)}
}
\description{
Facet for relative abundance data
}
