% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-geoms.R
\docType{data}
\name{geom_col_segsh}
\alias{geom_col_segsh}
\alias{geom_col_segs}
\alias{GeomColSegsh}
\alias{GeomColSegs}
\alias{GeomLineh}
\title{Useful geometries for strat diagrams}
\format{
An object of class \code{GeomColSegsh} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomColSegs} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomLineh} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
geom_col_segsh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  xend = 0,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_col_segs(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  yend = 0,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

GeomColSegsh

GeomColSegs

GeomLineh
}
\arguments{
\item{mapping, data, stat, position, arrow, arrow.fill, lineend, linejoin, na.rm, show.legend, inherit.aes, ...}{See
\link[ggplot2:geom_segment]{geom_segment}.}

\item{xend, yend}{The end of the horizontal or vertical segment bars, respectively.}
}
\value{
A ggplot2 layer
}
\description{
Useful geometries for strat diagrams
}
\keyword{datasets}
