% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat.R
\name{geom_colh}
\alias{geom_colh}
\title{Horizontal bar geometry}
\usage{
geom_colh(mapping = NULL, data = NULL, position = "dodgev", ...,
  width = NULL, na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping, data, position, ..., width, na.rm, show.legend, inherit.aes}{See
\link[ggstance]{geom_colh}}
}
\value{
A ggplot2 layer
}
\description{
Differs from ggstance::\link[ggstance]{geom_colh} by setting the default position
to dodge rather than stack, which is more likely to be useful in a stratigraphic context
}
