% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat.R
\docType{data}
\name{geom_hsegments}
\alias{geom_hsegments}
\alias{geom_vsegments}
\alias{GeomHsegments}
\alias{GeomVsegments}
\title{Useful geometries for strat diagrams}
\format{An object of class \code{GeomHsegments} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
geom_hsegments(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., xend = 0, arrow = NULL, arrow.fill = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_vsegments(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., yend = 0, arrow = NULL, arrow.fill = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

GeomHsegments

GeomVsegments
}
\arguments{
\item{mapping, data, stat, position, arrow, arrow.fill, lineend, linejoin, na.rm, show.legend, inherit.aes, ...}{See
\link[ggplot2]{geom_segment}.}

\item{xend, yend}{The end of the horizontal or vertical segment bars, respectively.}
}
\value{
A ggplot2 layer
}
\description{
Useful geometries for strat diagrams
}
\keyword{datasets}
