% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-geoms.R
\name{geom_lineh}
\alias{geom_lineh}
\title{Connect observations in the vertical direction}
\usage{
geom_lineh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{See
\link[ggplot2:geom_path]{geom_line}.}
}
\value{
A ggplot2 layer.
}
\description{
Connect observations in the vertical direction
}
