% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-themes.R
\name{ggplot_add.paleo_hook}
\alias{ggplot_add.paleo_hook}
\title{Infrastructure for special methods}
\usage{
\method{ggplot_add}{paleo_hook}(object, plot, object_name)
}
\arguments{
\item{object, object_name}{ggplot_add arguments}

\item{plot}{A ggplot object}
}
\value{
A ggplot object
}
\description{
These methods provide the infrastructure necessary to rotate facet labels and set
multiple scales.
}
