% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halifax_lakes}
\alias{halifax_lakes}
\title{Halifax lakes water chemistry and top/bottom diatom counts}
\format{An object of class \code{mudata} (inherits from \code{list}) of length 5.}
\source{
Neotoma paleoecology database (\url{http://neotomadb.org/})
}
\usage{
halifax_lakes
}
\description{
A subset of well-labeled, clean diatom count data for 44 Halifax-area (Nova Scotia) lakes,
an analysis of which has been published by Ginn et al. (2015). These data are
in the form of a \link[mudata2]{mudata} object, from which location
information, parameter information, and data can be extracted using
\link[mudata2]{tbl_locations}, \link[mudata2]{tbl_params}, and
\link[mudata2]{tbl_data} (respectively).
}
\examples{
library(mudata2)
data(halifax_lakes)

halifax_lakes \%>\%
  tbl_data()

}
\references{
Ginn, Brian K., Thiyake Rajaratnam, Brian F. Cumming, and John P.
Smol. "Establishing Realistic Management Objectives for Urban Lakes Using
Paleolimnological Techniques: An Example from Halifax Region (Nova Scotia,
Canada)." Lake and Reservoir Management 31, no. 2 (April 3, 2015): 92-108.
\url{https://doi.org/10.1080/10402381.2015.1013648}.
}
\keyword{datasets}
