% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halifax_lakes_plottable}
\alias{halifax_lakes_plottable}
\title{Halifax lakes water chemistry and top/bottom diatom counts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 5 columns.
}
\source{
Neotoma paleoecology database (\url{https://neotomadb.org})
}
\usage{
halifax_lakes_plottable
}
\description{
A subset of well-labeled, clean diatom count data for 44 Halifax-area (Nova Scotia) lakes,
an analysis of which has been published by Ginn et al. (2015).
}
\examples{
halifax_lakes_plottable

}
\references{
Ginn, Brian K., Thiyake Rajaratnam, Brian F. Cumming, and John P.
Smol. "Establishing Realistic Management Objectives for Urban Lakes Using
Paleolimnological Techniques: An Example from Halifax Region (Nova Scotia,
Canada)." Lake and Reservoir Management 31, no. 2 (April 3, 2015): 92-108.
\doi{10.1080/10402381.2015.1013648}.
}
\keyword{datasets}
