% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{is_age_depth_model}
\alias{is_age_depth_model}
\title{Test for age depth models}
\usage{
is_age_depth_model(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object is an age depth model, FALSE otherwise
}
\description{
Test for age depth models
}
