% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keji_lakes}
\alias{keji_lakes}
\alias{keji_lakes_plottable}
\title{Keji lakes core diatom counts}
\format{An object of class \code{mudata} (inherits from \code{list}) of length 5.}
\source{
Neotoma paleoecology database (\url{http://neotomadb.org/})
}
\usage{
keji_lakes

keji_lakes_plottable
}
\description{
A subset of well-labeled, clean diatom count data for 3 Keji-area (Nova
Scotia) lakes, which form part of the analysis in Ginn et al. (2007).
These data are in the form of a \link[mudata2]{mudata} object, from which
location information, parameter information, and data can be extracted using
\link[mudata2]{tbl_locations}, \link[mudata2]{tbl_params}, and
\link[mudata2]{tbl_data} (respectively). Also, a subset of these
data amenable to plotting have been made available.
}
\examples{
library(mudata2)
data(keji_lakes)

keji_lakes \%>\%
  tbl_data()

}
\references{
Ginn, Brian K., Brian F. Cumming, and John P. Smol. "Long-Term
Lake Acidification Trends in High- and Low-Sulphate Deposition Regions from
Nova Scotia, Canada." Hydrobiologia 586, no. 1 (July 1, 2007): 261-75.
\url{https://doi.org/10.1007/s10750-007-0644-3}.
}
\keyword{datasets}
