% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keji_lakes_plottable}
\alias{keji_lakes_plottable}
\title{Keji lakes core diatom counts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 202 rows and 5 columns.
}
\source{
Neotoma paleoecology database (\url{https://www.neotomadb.org})
}
\usage{
keji_lakes_plottable
}
\description{
A subset of well-labeled, clean diatom count data for 3 Keji-area (Nova
Scotia) lakes, which form part of the analysis in Ginn et al. (2007).
}
\examples{
keji_lakes_plottable

}
\references{
Ginn, Brian K., Brian F. Cumming, and John P. Smol. "Long-Term
Lake Acidification Trends in High- and Low-Sulphate Deposition Regions from
Nova Scotia, Canada." Hydrobiologia 586, no. 1 (July 1, 2007): 261-75.
\doi{10.1007/s10750-007-0644-3}.
}
\keyword{datasets}
