% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kellys_lake}
\alias{kellys_lake}
\alias{kellys_lake_cladocera}
\alias{kellys_lake_geochem}
\alias{kellys_lake_ages}
\title{Kellys Lake Data}
\format{An object of class \code{mudata} (inherits from \code{list}) of length 5.}
\usage{
kellys_lake

kellys_lake_cladocera

kellys_lake_geochem

kellys_lake_ages
}
\description{
Geochemistry measurements and Cladocera counts from Kellys Lake,
Cape Breton Island, Nova Scotia, Canada.
}
\references{
Joshua Kurek, Ian Spooner, and Dewey Dunnington (unpublished data).
}
\keyword{datasets}
