% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat.R
\name{label_species}
\alias{label_species}
\title{Facet labellers}
\usage{
label_species(labels, dont_italicize = c("\\\\(.*?\\\\)", "spp?\\\\.",
  "-complex"), species_facet = 1, multi_line = TRUE)
}
\arguments{
\item{labels}{A data.frame of facet label values}

\item{dont_italicize}{Regular expressions that should not be italicized}

\item{species_facet}{Which facet(s) contain species values}

\item{multi_line}{See \link[ggplot2]{label_parsed}}
}
\description{
Use these to label species with partial italic formatting. See \link[ggplot2]{label_parsed}.
}
