% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-plot_addons.R
\name{layer_dendrogram}
\alias{layer_dendrogram}
\alias{layer_zone_boundaries}
\title{Add a dendrogram as a layer or facet}
\usage{
layer_dendrogram(object, mapping, ...,
  segment_geom = ggplot2::geom_segment(inherit.aes = FALSE),
  sequential_facets = TRUE)

layer_zone_boundaries(object, mapping, ..., lty = 2, alpha = 0.7,
  colour = "black", size = 0.5)
}
\arguments{
\item{object}{A \link{nested_hclust} object.}

\item{mapping}{Map at least one axis (x or y) to a qualifier, like \code{aes(x = depth)} or similar.}

\item{...}{Use facet_var = "CONISS" or similar to name the panel}

\item{segment_geom}{Pass some variant of \link[ggplot2]{geom_segment} to customize geometry options}

\item{sequential_facets}{TRUE will result in the panel containing the dendrogram added to the right
of the plot.}

\item{lty, alpha, colour, size}{Customize the apperance of boundary lines}
}
\description{
Add a dendrogram as a layer or facet
}
