% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_chclust.R
\name{nested_chclust}
\alias{nested_chclust}
\title{Nested Constrained hierarchical clustering (CONISS)}
\usage{
nested_chclust(.data, data_column = "data",
  qualifiers_column = "qualifiers", distance_fun = stats::dist,
  n_groups = NULL, ...)
}
\arguments{
\item{.data}{A data frame with a list column of data frames, possibly created using
\link{nested_data_matrix}.}

\item{data_column}{The name of the column that contains the data. Evaluated
like \link[dplyr]{pull}.}

\item{qualifiers_column}{The column that contains the qualifiers}

\item{distance_fun}{A distance function like \link[stats]{dist} or \link[vegan]{vegdist}.}

\item{n_groups}{The number of groups to use (can be a vector or expression using vars in .data)}

\item{...}{Passed to \link[rioja]{chclust}.}
}
\value{
.data with additional columns 'model', 'broken_stick', and 'segments'
}
\description{
Powered by \link[rioja]{chclust}; broken stick using \link[rioja]{bstick}.
}
\examples{
library(tidyr)
library(dplyr)

nested_coniss <- keji_lakes_plottable \%>\%
  group_by(location) \%>\%
  nested_data_matrix(taxon, rel_abund, depth, fill = 0) \%>\%
  nested_chclust()

# plot broken stick dispersion to verify number of plausible groups
library(ggplot2)

nested_coniss \%>\%
  unnest(broken_stick) \%>\%
  tidyr::gather(type, value, broken_stick_dispersion, dispersion) \%>\%
  ggplot(aes(x = n_groups, y = value, col = type)) +
  geom_line() +
  geom_point() +
  facet_wrap(vars(location))

}
\references{
Grimm, E.C. (1987) CONISS: A FORTRAN 77 program for stratigraphically constrained cluster
analysis by the method of incremental sum of squares. Computers & Geosciences, 13, 13-35.
\url{http://doi.org/10.1016/0098-3004(87)90022-7}

Juggins, S. (2017) rioja: Analysis of Quaternary Science Data, R package version (0.9-15.1).
(\url{http://cran.r-project.org/package=rioja}).
}
