% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_prcomp.R
\name{nested_prcomp}
\alias{nested_prcomp}
\title{Nested Principal Components Analysis (PCA)}
\usage{
nested_prcomp(.data, data_column = "data", ...)
}
\arguments{
\item{.data}{A data frame with a list column of data frames, possibly created using
\link{nested_data_matrix}.}

\item{data_column}{The name of the column that contains the data. Evaluated
like \link[dplyr]{pull}.}

\item{...}{Passed to \link[stats]{prcomp}.}
}
\value{
.data with additional columns 'model', 'loadings', 'variance' and 'scores'
}
\description{
Powered by \link[stats]{prcomp}. When creating the \link{nested_data_matrix},
the data should be scaled (i.e, \code{trans = scale}) if all variables are not
in the same unit.
}
\examples{
library(tidyr)

nested_pca <- alta_lake_geochem \%>\%
  nested_data_matrix(
    key = param,
    value = value,
    qualifiers = c(depth, zone),
    trans = scale
  ) \%>\%
  nested_prcomp()

# get variance info
nested_pca \%>\% unnest(variance)

# get loadings info
nested_pca \%>\% unnest(loadings)

# scores, requalified
nested_pca \%>\% unnest(qualifiers, scores)

}
