% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{new_age_depth_model}
\alias{new_age_depth_model}
\title{Create age depth model objects}
\usage{
new_age_depth_model(x)
}
\arguments{
\item{x}{A list to be classed as an age depth model object}
}
\value{
An age depth model object
}
\description{
Create age depth model objects
}
