% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data_matrix.R, R/nested_prcomp.R
\name{plot.nested_analysis}
\alias{plot.nested_analysis}
\alias{plot_nested_analysis}
\alias{biplot.nested_prcomp}
\title{Plot a nested analysis}
\usage{
\method{plot}{nested_analysis}(x, ..., main = "", nrow = NULL,
  ncol = NULL)

plot_nested_analysis(.x, .fun, ..., nrow = NULL, ncol = NULL)

\method{biplot}{nested_prcomp}(x, ..., nrow = NULL, ncol = NULL)
}
\arguments{
\item{x, .x}{A \link{nested_analysis} object (or subclass)}

\item{...}{Passed to the plot function. Tidy evaluation is supported, and arguments are evaluated
within a transposed version of x for each row.}

\item{main}{The plot title}

\item{nrow, ncol}{Force a number of rows or columns in the output}

\item{.fun}{A function that produces graphical output}
}
\value{
A list containing the result of the plot function (invisibly)
}
\description{
Calls \link[graphics]{plot} or another (base) plotting function on all models, arranging the output in subplots.
}
