% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{print.age_depth_model}
\alias{print.age_depth_model}
\title{Print summary of age-depth model objects}
\usage{
\method{print}{age_depth_model}(x, ...)
}
\arguments{
\item{x}{An \link{age_depth_model}}

\item{...}{Unused}
}
\value{
The input, invisibly
}
\description{
Print summary of age-depth model objects
}
