% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_paleo.R
\name{rotate_facet_labels}
\alias{rotate_facet_labels}
\alias{remove_label_clip}
\title{Common plot modifications for stratigraphic plots}
\usage{
rotate_facet_labels(angle = 45, direction = "x",
  remove_label_background = TRUE)

remove_label_clip(direction)
}
\arguments{
\item{angle}{The angle at which labels should be rotated}

\item{direction}{The axes along which the operations should be performed}

\item{remove_label_background}{Whether or not label backgrounds should be removed along
rotated label axes}
}
\value{
An object or list of objects that can be added to a \link[ggplot2]{ggplot}
}
\description{
Common plot modifications for stratigraphic plots
}
