% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-themes.R
\name{theme_paleo}
\alias{theme_paleo}
\title{A Paleo-friendly ggplot2 theme}
\usage{
theme_paleo(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}}
}
\value{
A complete \code{\link[ggplot2:theme]{ggplot2::theme()}}
}
\description{
Essentially, this is \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}} with a few modifications
}
\examples{
library(ggplot2)

ggplot(mpg, aes(cty, hwy)) +
  geom_point() +
  theme_paleo()

}
