% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypaleo-package.R
\docType{package}
\name{tidypaleo-package}
\alias{tidypaleo}
\alias{tidypaleo-package}
\title{tidypaleo: Tidy tools for paleoenvironmental archives}
\description{
Provides a set of functions with a common framework for age-depth model management, 
  stratigraphic visualization, and common statistical transformations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/tidypaleo}
  \item \url{https://github.com/paleolimbot/tidypaleo}
  \item Report bugs at \url{https://github.com/paleolimbot/tidypaleo/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net}

}
\keyword{internal}
