% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{trans_interpolate}
\alias{trans_interpolate}
\alias{trans_average}
\alias{trans_exact}
\alias{trans_na}
\title{Transforms}
\usage{
trans_interpolate(x, y)

trans_average(x, y, x0 = last, y0 = last, slope = NULL)

trans_exact(x, y)

trans_na(x, y)
}
\arguments{
\item{x}{A paired vector of x values}

\item{y}{A paired vector of y values}

\item{x0}{The x value to anchor the transform}

\item{y0}{The y value to anchor the transform}

\item{slope}{The slope (in units of y/x) to use for the transform}
}
\value{
A list with component functions \code{trans} and \code{inverse}
}
\description{
Transforms
}
