% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R
\name{validate_age_depth_model}
\alias{validate_age_depth_model}
\title{Validate age depth model objects}
\usage{
validate_age_depth_model(x)
}
\arguments{
\item{x}{An age depth model object}
}
\value{
The input, invisibly
}
\description{
Validate age depth model objects
}
