% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat.R
\name{facet_abundanceh}
\alias{facet_abundanceh}
\alias{facet_abundance}
\title{Facet for relative abundance data}
\usage{
facet_abundanceh(taxon, grouping = NULL, rotate_facet_labels = 45,
  labeller = label_species, scales = "free_x", space = "free_x", ...)

facet_abundance(taxon, grouping = NULL, rotate_facet_labels = 0,
  labeller = label_species, scales = "free_y", space = "free_y", ...)
}
\arguments{
\item{taxon}{A call to \link[ggplot2]{vars}, defining the column that identifies the taxon.}

\item{grouping}{A call to \link[ggplot2]{vars}, identifying additional grouping columns}

\item{rotate_facet_labels}{Facet label rotation (degrees)}

\item{labeller}{Labeller to process facet names. Use \link{label_species} to italicize
species names, or \link[ggplot2]{label_value} to suppress.}

\item{space, scales}{Modify default scale freedom behaviour}

\item{...}{Passed to \link[ggplot2]{facet_grid}.}
}
\description{
Provides a number of modifications to the plot that are necessary for relative abundance plots
of a number of species. See \link{scale_x_abundance}, \link[ggplot2]{facet_grid},
and \link{rotated_facet_labels} for examples of how to customize the default behaviour.
}
