% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_lake_14C_ages}
\alias{long_lake_14C_ages}
\alias{long_lake_bacon_ages}
\alias{long_lake_plottable}
\title{Long Lake Carbon-14 Ages}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 86 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 10 columns.
}
\usage{
long_lake_14C_ages

long_lake_bacon_ages

long_lake_plottable
}
\description{
This object contains several uncalibrated Carbon-14 measurements from Long
Lake (Nova Scotia-New Brunswick Border Region, Canada) core LL-PC2
(Dunnington et al. 2017; White 2012). The \code{long_lake_bacon_ages} object
contains the result of the Carbon-14 ages as
modelled by the rbacon package (Blaauw and Christen 2011).
}
\examples{
long_lake_14C_ages
long_lake_bacon_ages
long_lake_plottable

}
\references{
Blaauw, Maarten, and J. Andrés Christen. "Flexible Paleoclimate Age-Depth
Models Using an Autoregressive Gamma Process." Bayesian Analysis 6, no. 3
(September 2011): 457–74. \doi{10.1214/ba/1339616472}.

Dunnington, Dewey W., Hilary White, Ian S. Spooner, Mark L. Mallory, Chris
White, Nelson J. O’Driscoll, and Nic R. McLellan. "A Paleolimnological
Archive of Metal Sequestration and Release in the Cumberland Basin Marshes,
Atlantic Canada." FACETS 2, no. 1 (May 23, 2017): 440–60.
\doi{10.1139/facets-2017-0004}.

White, Hilary E. "Paleolimnological Records of Post-Glacial Lake
and Wetland Evolution from the Isthmus of Chignecto Region, Eastern Canada."
M.Sc. Thesis, Acadia University, 2012.
\url{https://scholar.acadiau.ca/islandora/object/theses:247}.
}
\keyword{datasets}
