% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data_matrix.R
\name{nested_anal}
\alias{nested_anal}
\alias{nested_pca}
\title{Perform an analysis on a nested data matrix}
\usage{
nested_anal(.data, data_column, fun, data_arg, ..., model_column = "model")

nested_pca(.data, data_column = "data", ...)
}
\arguments{
\item{.data}{A data frame with a list column of data frames}

\item{data_column}{The column that contains the data}

\item{fun}{A model function}

\item{data_arg}{The data argument of fun}

\item{...}{Passed to fun}

\item{model_column}{The column that will contain the model(s)}
}
\value{
.data with an additional list column of fun output
}
\description{
Perform an analysis on a nested data matrix
}
