% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data_matrix.R
\name{nested_anal}
\alias{nested_anal}
\title{Perform an analysis on a nested data matrix}
\usage{
nested_anal(.data, data_column, fun, data_arg, ..., model_column = "model",
  reserved_names = NULL)
}
\arguments{
\item{.data}{A data frame with a list column of data frames, possibly created using
\link{nested_data_matrix}.}

\item{data_column}{The name of the column that contains the data. Evaluated
like \link[dplyr]{pull}.}

\item{fun}{A model function}

\item{data_arg}{The data argument of fun}

\item{...}{Passed to fun}

\item{model_column}{The column that will contain the model(s)}

\item{reserved_names}{Names that should not be allowed as columns in any
data frame within this object}
}
\value{
.data with an additional list column of fun output
}
\description{
Perform an analysis on a nested data matrix
}
