% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_data_matrix.R
\name{nested_chclust}
\alias{nested_chclust}
\title{Nested Constrained hierarchical clustering (CONISS)}
\usage{
nested_chclust(.data, data_column = "data",
  qualifiers_column = "qualifiers", distance_fun = stats::dist,
  n_groups = NULL, ...)
}
\arguments{
\item{.data}{A data frame with a list column of data frames, possibly created using
\link{nested_data_matrix}.}

\item{data_column}{The name of the column that contains the data. Evaluated
like \link[dplyr]{pull}.}

\item{distance_fun}{A distance function like \link[stats]{dist} or \link[vegan]{vegdist}.}

\item{...}{Passed to \link[rioja]{chclust}.}
}
\value{
.data with additional columns 'model', 'broken_stick', and 'segments'
}
\description{
Powered by \link[rioja]{chclust}; broken stick using \link[rioja]{bstick}.
}
\references{
Grimm, E.C. (1987) CONISS: A FORTRAN 77 program for stratigraphically constrained cluster
analysis by the method of incremental sum of squares. Computers & Geosciences, 13, 13-35.
\url{http://doi.org/10.1016/0098-3004(87)90022-7}

Juggins, S. (2017) rioja: Analysis of Quaternary Science Data, R package version (0.9-15.1).
(\url{http://cran.r-project.org/package=rioja}).
}
