% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrat-facets.R, R/ggstrat-geoms.R,
%   R/nested_data_matrix.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vars}
\alias{position_fillv}
\alias{PositionFillv}
\alias{position_stackv}
\alias{PositionStackv}
\alias{position_dodge2v}
\alias{PositionDodge2v}
\alias{position_dodgev}
\alias{PositionDodgev}
\alias{geom_colh}
\alias{GeomColh}
\alias{filter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{ggplot2}{\code{\link[ggplot2]{vars}}}

  \item{ggstance}{\code{\link[ggstance]{geom_colh}}, \code{\link[ggstance]{GeomColh}}, \code{\link[ggstance]{position_dodge2v}}, \code{\link[ggstance]{position_dodgev}}, \code{\link[ggstance]{position_fillv}}, \code{\link[ggstance]{position_stackv}}, \code{\link[ggstance]{PositionDodge2v}}, \code{\link[ggstance]{PositionDodgev}}, \code{\link[ggstance]{PositionFillv}}, \code{\link[ggstance]{PositionStackv}}}
}}

