% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_depth_model.R, R/ggstrat-facets.R,
%   R/ggstrat-geoms.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{first}
\alias{last}
\alias{vars}
\alias{position_fillv}
\alias{PositionFillv}
\alias{position_stackv}
\alias{PositionStackv}
\alias{position_dodge2v}
\alias{PositionDodge2v}
\alias{position_dodgev}
\alias{PositionDodgev}
\alias{geom_colh}
\alias{GeomColh}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:nth]{first}}, \code{\link[dplyr:nth]{last}}}

  \item{ggplot2}{\code{\link[ggplot2]{vars}}}

  \item{ggstance}{\code{\link[ggstance:ggstance-ggproto]{GeomColh}}, \code{\link[ggstance:ggstance-ggproto]{PositionDodge2v}}, \code{\link[ggstance:ggstance-ggproto]{PositionDodgev}}, \code{\link[ggstance:ggstance-ggproto]{PositionFillv}}, \code{\link[ggstance:ggstance-ggproto]{PositionStackv}}, \code{\link[ggstance:geom_barh]{geom_colh}}, \code{\link[ggstance:position-vertical]{position_dodge2v}}, \code{\link[ggstance:position-vertical]{position_dodgev}}, \code{\link[ggstance:position-vertical]{position_fillv}}, \code{\link[ggstance:position-vertical]{position_stackv}}}
}}

