% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypaleo-package.R
\docType{package}
\name{tidypaleo-package}
\alias{tidypaleo}
\alias{tidypaleo-package}
\title{tidypaleo: Tidy Tools for Paleoenvironmental Archives}
\description{
Provides a set of functions with a common framework for age-depth model management, stratigraphic visualization, and common statistical transformations. The focus of the package is stratigraphic visualization, for which 'ggplot2' components are provided to reproduce the scales, geometries, facets, and theme elements commonly used in publication-quality stratigraphic diagrams. Helpers are also provided to reproduce the exploratory statistical summaries that are frequently included on stratigraphic diagrams. See Dunnington et al. (2021) <doi:10.18637/jss.v101.i07>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/tidypaleo/}
  \item \url{https://github.com/paleolimbot/tidypaleo}
  \item Report bugs at \url{https://github.com/paleolimbot/tidypaleo/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID}) [copyright holder]

}
\keyword{internal}
